+++ internal/ssh/connect.go
@@ -207,7 +207,9 @@ func acceptAndSaveHostKey(path string) ssh.HostKeyCallback {
 }
 
 func saveHostKey(path, hostname string, key ssh.PublicKey) error {
-	_ = os.MkdirAll(filepath.Dir(path), 0700)
+	if err := os.MkdirAll(filepath.Dir(path), 0700); err != nil {
+		return err
+	}
 	f, err := os.OpenFile(path, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0600)
 	if err != nil {
 		return err